+++ app/src/main/java/com/pict/pbl/medswift/login/LoginManager.kt
@@ -2,6 +2,7 @@ package com.pict.pbl.medswift.login
 
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
+import com.pict.pbl.medswift.data.User
 import com.pict.pbl.medswift.viewmodels.LoginViewModel
 import java.util.regex.Pattern
 
@@ -24,10 +25,24 @@ class LoginManager( private val loginViewModel: LoginViewModel? = null ) {
 
     }
 
-    fun createUser( emailAddress: String , password: String ) {
-        firebaseAuth.createUserWithEmailAndPassword( emailAddress , password )
+    fun loginUser( emailAddress: String , password: String , result : ( () -> Unit ) ) {
+        firebaseAuth.signInWithEmailAndPassword( emailAddress , password )
             .addOnSuccessListener {
-                // TODO: Send verification email
+                result()
+            }
+            .addOnFailureListener {
+                if( loginViewModel != null ) {
+                    loginViewModel.errorMessage.value = it.message
+                    loginViewModel.errorMessageFlag.value = true
+                }
+            }
+    }
+
+
+    fun createUser( user: User , password: String ) {
+        firebaseAuth.createUserWithEmailAndPassword( user.email , password )
+            .addOnSuccessListener {
+                // TODO: Send verification email and create user in DB
             }
             .addOnFailureListener {
                 if( loginViewModel != null ) {