+++ app/src/main/java/com/pict/pbl/medswift/screens/history/History.kt
@@ -1,0 +1,26 @@
+package com.pict.pbl.medswift.screens.history
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
+
+@Composable
+fun HistoryScreen() {
+    MedSwiftTheme {
+        Surface(modifier = Modifier
+            .fillMaxSize()
+            .background(Color.White)) {
+            ScreenUI()
+        }
+    }
+}
+
+@Composable
+private fun ScreenUI() {
+
+}
+